requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.7.0
anytree==2.12.1
appnope==0.1.4
APScheduler==3.11.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
awscli==1.44.54
azure-core==1.39.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.42.73
botocore==1.42.73
casadi==3.6.7
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.0
claude-code-proxy==0.1.5
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
croniter==6.2.2
cryptography==43.0.3
cycler==0.12.1
ddgs==9.12.0
deap==1.4.3
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docutils==0.19
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastapi-sso==0.16.0
fastar==0.9.0
fastjsonschema==2.21.1
fastuuid==0.14.0
filelock==3.17.0
Flask==3.1.3
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hf-xet==1.4.2
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
httpx-aiohttp==0.1.12
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.10
imageio==2.36.1
importlib_metadata==9.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.13.0
jmespath==1.1.0
joblib==1.4.2
json5==0.10.0
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
keras==3.7.0
kiwisolver==1.4.7
lab==8.3
latexcodec==3.0.0
lazy_loader==0.4
libclang==18.1.1
litellm==1.82.5
litellm-enterprise==0.1.34
litellm-proxy-extras==0.4.58
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mcp==1.26.0
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.0.2
ml-dtypes==0.4.1
monotonic==1.6
mpmath==1.3.0
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openai==2.24.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
polars==1.39.3
polars-runtime-32==1.39.3
pooch==1.8.2
posthog==3.7.5
primp==1.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.4.1
prosperity3bt==0.7.1
protobuf==4.25.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pybamm==25.1.1
pybammsolvers==0.0.5
pybtex==0.24.0
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.12.1
PyNaCl==1.6.2
pyparsing==3.2.0
pyroscope-io==0.8.16
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.22
python-slugify==8.0.4
python-utils==3.9.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdp==0.8
redis==7.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.22.3
rq==2.7.0
rsa==4.7.2
s3transfer==0.16.0
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.55.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.46.2
svgwrite==1.1.6
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf-keras-vis==0.8.7
threadpoolctl==3.5.0
tifffile==2025.5.10
tiktoken==0.12.0
tinycss2==1.4.0
together==1.5.29
tokenizers==0.22.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
txt2tags==3.9
typer==0.24.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.42.0
uvloop==0.21.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xarray==2025.1.0
yarl==1.22.0
zipp==3.23.0
